feat: use func to adjust session state
