+++ internal/handlers/handler_sign_u2f_step2.go
@@ -2,9 +2,7 @@ package handlers
 
 import (
 	"fmt"
-	"time"
 
-	"github.com/authelia/authelia/internal/authentication"
 	"github.com/authelia/authelia/internal/middlewares"
 )
 
@@ -48,10 +46,7 @@ func SecondFactorU2FSignPost(u2fVerifier U2FVerifier) middlewares.RequestHandler
 			return
 		}
 
-		userSession.AuthenticationLevel = authentication.TwoFactor
-		if userSession.SecondFactorAuthn == 0 {
-			userSession.SecondFactorAuthn = time.Now().Unix()
-		}
+		userSession.SetTwoFactor(ctx.Clock.Now())
 
 		err = ctx.SaveSession(userSession)
 		if err != nil {
